% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-table-mosaic.R
\name{autoplot_table_mosaic}
\alias{autoplot_table_mosaic}
\title{Autoplot table mosaic}
\usage{
autoplot_table_mosaic(x)
}
\arguments{
\item{x}{A table}
}
\value{
a `ggplot2` object
}
\description{
Create a `ggplot2` object from a `table` with `ggmosaic`
}
