% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\title{cache an object}
\usage{
cache(name, expr, overwrite = FALSE, time = 3600, env = parent.frame())
}
\arguments{
\item{name}{Name of the object}

\item{expr}{Expression to evaluate}

\item{overwrite}{If `TRUE` saved over current objects}

\item{time}{Either time (seconds) for how long to cache the object(s) or the
time as a `POSIXt` object}

\item{env}{Environment to load object into if found}
}
\value{
Result of `expr`, or cached result
}
\description{
cache an R object
}
\examples{
res <- .POSIXct(0)
for (i in 1:6) {
  cache("foo", Sys.time(), time = 1)
  Sys.sleep(0.5)
  res[i] <- foo
}
res
}
