% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-equation.R
\name{coef_equation}
\alias{coef_equation}
\title{Print a model as an equation}
\usage{
coef_equation(mod, FUN = sf2, ...)
}
\arguments{
\item{mod}{A model, something which `stats::coef()` can extracted a named
vector}

\item{FUN}{A function to apply to the values of `stats::coef(mod)`, for
formatting; default formats to 2 significant figures with trailing zeros}

\item{...}{Additional arguments passed to `FUN`}
}
\description{
Print a model as an equation
}
\examples{
mod <- lm(Sepal.Length ~ Sepal.Width + Petal.Length, data = iris)
coef_equation(mod)
# "2.2 + Sepal.Width \* 0.60 + Petal.Length \* 0.47"
}
