% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{default_env}
\alias{default_env}
\title{Default environment}
\usage{
default_env(key = "default")
}
\arguments{
\item{key}{A key for which environments to use}
}
\description{
Creates an environment with just default packages
}
\details{
A new environment is created with the parent of
}
