% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-combine.R
\name{filter_combine}
\alias{filter_combine}
\title{Filter combine}
\usage{
filter_combine(x, ..., .id, .names)
}
\arguments{
\item{x}{A data.frame}

\item{...}{
  Arguments passed on to \code{\link[dplyr:filter]{dplyr::filter}}
  \describe{
    \item{\code{.by}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Optionally, a selection of columns to
group by for just this operation, functioning as an alternative to \code{\link[dplyr:group_by]{group_by()}}. For
details and examples, see \link[dplyr:dplyr_by]{?dplyr_by}.}
  }}

\item{.id}{When .id is supplied, a new column of identifiers is created to
link each row to its original data frame. The labels are taken from
`.names`}

\item{.names}{vector of characters for the}
}
\description{
Filters a data frame then binds the result with it (with names).
}
\seealso{
[dplyr::filter()] and [dplyr::bind_rows()]
}
