% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fishers_method}
\alias{fishers_method}
\title{Fisher's method for combined probabilities}
\usage{
fishers_method(x, zeros = FALSE)
}
\arguments{
\item{x}{A vector of p-values.}

\item{zeros}{Logical, if `TRUE` will use zeros in calculation, otherwise
recoded as `.Machine$double.xmin`.  This still will produce a very small
result but will be less likely to produce `0`}
}
\description{
Computes Fisher's method for combined probabilities
}
\details{
Values greater than 1 and less than 0 are removed from the calculation.
  Values of 0 are red
}
