% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newer.R
\name{\%newer\%}
\alias{\%newer\%}
\title{Is newer}
\usage{
x \%newer\% y
}
\arguments{
\item{x, y, }{Files (as character paths) or `Date`/`POSIXt` object}
}
\value{
`TRUE` or `FALSE` for if `x` is newer than `y`.  If file `y` does
  not exist, `TRUE` is returned
}
\description{
Check if a file or time is newer than another
}
