% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jordanExtra.R
\docType{package}
\name{jordanExtra}
\alias{jordanExtra}
\alias{jordanExtra-package}
\title{jordanExtra}
\description{
Extra functions that may not be generalizable enough for `jordan`
}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
