% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-password.R
\name{make_password}
\alias{make_password}
\title{Generate a random password}
\usage{
make_password(n = 16L, specials = "!@#$\%^&*_+-=?", silent = TRUE)
}
\arguments{
\item{n}{The character length desired (default: `16L`)}

\item{specials}{A single string of characters other than all letterrs (from
`LETTERS` and `letters`) and numbers (`0:9`)}

\item{silent}{if `FALSE` will call `cat()` on the password}
}
\value{
`NULL`, invisible
}
\description{
Make a password and send to the clipboard with [jordan::write_clipboard()]
}
