% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markExtra-package.R
\docType{package}
\name{markExtra-package}
\alias{markExtra}
\alias{markExtra-package}
\title{markExtra: Extra Functions for the mark Package}
\description{
Extra functions for the mark package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jmbarbone.github.io/markExtra/}
}

}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
