% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min-max.R
\name{min_max}
\alias{min_max}
\title{Min-max normalization}
\usage{
min_max(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{A vector of `numeric` values}

\item{lower, upper}{The boundaries of the normalized score; defaults to `0`
and `1`, respectively.  `upper` must be greater than `lower`.}
}
\value{
A vector of x normalizes so the lower scores is `lower` and highest
  score is `upper`
}
\description{
Min-max normalization
}
\examples{
x <- c(0.34, 0.24, 0.65, 0.21, 0.18, 0.79, 0.13, 0.94, 0.22, 0.1)
min_max(x)
min_max(x, upper = 10)
min_max(x, lower = -1)
min_max(x, l = 10, u = 20)
}
