% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passcode.R
\name{passcode}
\alias{passcode}
\title{passcode}
\usage{
passcode(
  len = 14:16,
  sep = "_",
  one = c("adjectives", "colors"),
  two = c("animals", "nouns"),
  numbers = len\%/\%4L
)
}
\arguments{
\item{len}{length of passcode}

\item{sep}{separator between words}

\item{one, two}{Type of words to use (see `?codename::codename` for `type`)}

\item{numbers}{Number of digits to append}
}
\value{
A passcode, as a string
}
\description{
Generate a passcode
}
\details{
See `?codename::codename` for more details
}
