% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{percentile_rank}
\alias{percentile_rank}
\title{Percentile rank}
\usage{
percentile_rank(x)
}
\arguments{
\item{x}{A vector of values to rank}
}
\value{
A percentile rank between 0 and 100
}
\description{
The bounds of a percentile rank are > 0 and < 100.

A percentile rank here is the proportion of scores that are less than the
  current score.

\deqn{PR = (c_L + 0.5 f_i) / N * 100}

Where

  \eqn{c_L} is the frequency of scores less than the score of interest

  \eqn{f_i} is the frequency of the score of interest
}
\details{
Computes a percentile rank for each score in a set.
}
\examples{
percentile_rank(0:9)
x <- c(1, 1, 2, 5, 7, NA_integer_, 7, 10)
percentile_rank(x)
# compare to dplyr
# dplyr::percent_rank(x) * 100
}
