% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{Reload}
\alias{Reload}
\alias{.Reload}
\alias{.Restart}
\alias{.CleanUp}
\title{Reload}
\usage{
.Reload(remove_objects = TRUE, loud = FALSE, env = parent.frame())

.Restart()

.CleanUp()
}
\arguments{
\item{remove_objects}{Logical, if `TRUE` will remove all objects found}

\item{loud}{Logical, if `TRUE` will present a message on removed objects}
}
\description{
Reloads your session
}
\seealso{
Other startup_utils: 
\code{\link{.NiceMessage}()},
\code{\link{attached_packages}},
\code{\link{startup_funs}}
}
\concept{startup_utils}
