% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\docType{data}
\name{startup_funs}
\alias{startup_funs}
\alias{.LoadFunctionsFromJordan}
\alias{.ResetOptions}
\alias{.GitBranchPrompt}
\alias{.LoadPipe}
\title{Start up functions}
\format{
An object of class \code{character} of length 8.
}
\usage{
.LoadFunctionsFromJordan(env = parent.frame())

.ResetOptions(keep_prompt = TRUE)

startup_funs

.GitBranchPrompt()

.LoadPipe(env = parent.frame())
}
\arguments{
\item{env}{The environment to save objects}

\item{keep_prompt}{Logical, if `FALSE` will not reset the prompt}
}
\description{
Functions that try to assist with start ups
}
\details{
These functions can be used to create an environment for startup.
}
\section{Start up}{

For more information on start up:
  * https://rstats.wtf/r-startup.html
  * `?Startup`
}

\seealso{
Other startup_utils: 
\code{\link{.NiceMessage}()},
\code{\link{Reload}},
\code{\link{attached_packages}}
}
\concept{startup_utils}
\keyword{datasets}
