% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{test_git_repo}
\alias{test_git_repo}
\title{Test a git repository}
\usage{
test_git_repo(x, ..., branch = "master", test_fun = devtools::test)
}
\arguments{
\item{x}{A project directory}

\item{...}{Additional arguments passed to test_fun}

\item{branch}{The branch to copy from the repository (default: "master")}

\item{test_fun}{A function to apply for testing (default: "devtools::test");
Note: the copied directory for testing is passed as the first argument}
}
\description{
Test a git repository
}
