% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markExtra.R
\docType{package}
\name{markExtra}
\alias{markExtra}
\alias{markExtra-package}
\title{markExtra}
\description{
Extra functions that may not be generalizable enough for `mark`
}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
