% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min-max.R
\name{min_max}
\alias{min_max}
\title{Min-max normalization}
\usage{
min_max(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{A vector of `numeric` values}

\item{lower, upper}{The boundaries of the normalized score; defaults to `0`
and `1`, respectively.  `upper` must be greater than `lower`.}
}
\value{
A vector of x normalizes so the lower scores is `lower` and highest score is `upper`
}
\description{
Min-max normalization
}
