% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-score.R
\name{weighted_score}
\alias{weighted_score}
\title{Weighted score}
\usage{
weighted_score(
  score,
  weights = 1,
  reference = mean(weights, na.rm = TRUE),
  comparison = mean(score, na.rm = TRUE)
)
}
\arguments{
\item{score}{A vector of values}

\item{weights}{Weights applied to each score}

\item{reference}{A reference value to compare against weights; when `weights`
are lower than `needed`, the final score is biased towards `population`.}

\item{comparison}{A comparison value; a population mean or other value}
}
\description{
Weight a score against a population
}
\examples{
scores <- rep(seq(0, 1, .1), each = 10)
weights <- rep(1:10, 10)
weighted <- weighted_score(scores, weights)
plot(weighted, xlab = "inded", ylab = "Weighted score")
plot(
  weighted_score(
    rep(seq(0, 1, .01))
  )
)
}
